_tda/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
appium-python-client==5.1.1
    # via -r requirements.in
attrs==25.3.0
    # via
    #   outcome
    #   trio
certifi==2025.7.9
    # via
    #   requests
    #   selenium
    #   sentry-sdk
charset-normalizer==3.4.2
    # via requests
execnet==2.1.1
    # via pytest-xdist
h11==0.16.0
    # via wsproto
idna==3.10
    # via
    #   requests
    #   trio
iniconfig==2.1.0
    # via pytest
outcome==1.3.0.post0
    # via
    #   trio
    #   trio-websocket
packaging==25.0
    # via pytest
pluggy==1.6.0
    # via pytest
py==1.11.0
    # via pytest-forked
pygments==2.19.2
    # via pytest
pysocks==1.7.1
    # via urllib3
pytest==8.4.1
    # via
    #   -r requirements.in
    #   pytest-forked
    #   pytest-xdist
pytest-forked==1.6.0
    # via -r requirements.in
pytest-timeout==2.4.0
    # via -r requirements.in
pytest-xdist==3.8.0
    # via -r requirements.in
pyyaml==6.0.2
    # via -r requirements.in
requests==2.32.4
    # via -r requirements.in
selenium==4.34.2
    # via
    #   -r requirements.in
    #   appium-python-client
sentry-sdk==2.32.0
    # via -r requirements.in
sniffio==1.3.1
    # via trio
sortedcontainers==2.4.0
    # via trio
trio==0.30.0
    # via
    #   selenium
    #   trio-websocket
trio-websocket==0.12.2
    # via selenium
typing-extensions==4.14.1
    # via selenium
urllib3[socks]==2.5.0
    # via
    #   -r requirements.in
    #   requests
    #   selenium
    #   sentry-sdk
websocket-client==1.8.0
    # via selenium
wsproto==1.2.0
    # via trio-websocket
